+++ CHANGES.md
@@ -18,6 +18,7 @@
   by dkml-base-compiler. <https://github.com/diskuv/dkml-compiler/issues/7>
 * The `lib/findlib.conf.d/<TARGETABI>` findlib configuration is generated for all cross-compilers.
 * Remove `-O0` from `/usr/bin/as` when `dkml-option-debuginfo` is present since `GNU as 2.30` does not support it.
+* Add -B option to only build bytecode to r-c-ocaml-1-setup.sh, and -3 option to only build 32-bit bytecode
 
 ## 2.1.3